Move 'transition-status' toggle to TransitionControl and add it to IB
